emergence_core/requirements.txt
numpy>=1.24.0
chromadb>=0.4.15
pydantic>=2.5.2
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
langchain>=0.0.325
sentence-transformers>=2.2.2
torch>=2.1.0
transformers>=4.34.0
accelerate>=0.28.0
bitsandbytes>=0.41.0
schedule>=1.2.0
discord.py>=2.3.2
aiohttp>=3.9.0
asyncio>=3.4.3
httpx>=0.24.1
arxiv>=1.4.7
wikipedia>=1.4.0
wolframalpha>=5.0.0
playwright>=1.39.0
docker>=6.1.3